added the fetch package
