+++ .babelrc
@@ -1,1 +1,7 @@
-{ "presets": ["@babel/preset-env"] }
+// { "presets": ["@babel/preset-env"] }
+{
+  "presets": ["@babel/preset-env"],
+  "plugins": [
+      ["@babel/transform-runtime"]
+  ]
+}
