+++ package.json
@@ -26,8 +26,13 @@
   "homepage": "https://github.com/Tresor11/Phaser-RGP-Game#readme",
   "devDependencies": {
     "@babel/core": "^7.9.0",
+    "@babel/plugin-transform-runtime": "^7.9.0",
     "@babel/preset-env": "^7.9.5",
+    "babel-core": "^6.26.3",
     "babel-loader": "^8.1.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-plugin-import": "^2.20.2",
@@ -39,7 +44,9 @@
     "webpack-dev-server": "^2.11.0"
   },
   "dependencies": {
+    "@babel/runtime": "^7.9.2",
     "http-server": "^0.12.1",
+    "node-fetch": "^2.6.0",
     "phaser": "^3.3.0"
   }
 }
